tools/transcription-saas/requirements.txt
fastapi
uvicorn
openai-whisper
yt-dlp
python-multipart
slowapi
